feat: update CodeMirror to ^6.0
