+++ apps/codeimage/package.json
@@ -36,41 +36,41 @@
     "@codeimage/highlight": "workspace:*",
     "@codeimage/dom-export": "workspace:*",
     "@codeimage/ui": "workspace:*",
-    "@codemirror/autocomplete": "^0.20.1",
-    "@codemirror/commands": "^0.20.0",
-    "@codemirror/lang-cpp": "^0.20.0",
-    "@codemirror/lang-css": "^0.20.0",
-    "@codemirror/lang-html": "^0.20.0",
-    "@codemirror/lang-java": "^0.20.0",
-    "@codemirror/lang-javascript": "^0.20.0",
-    "@codemirror/lang-json": "^0.20.0",
-    "@codemirror/lang-markdown": "^0.20.1",
-    "@codemirror/lang-php": "^0.20.0",
-    "@codemirror/lang-python": "^0.20.0",
-    "@codemirror/lang-rust": "^0.20.0",
-    "@codemirror/lang-sql": "^0.20.0",
-    "@codemirror/language": "^0.20.2",
-    "@codemirror/lint": "^0.20.2",
-    "@codemirror/search": "^0.20.1",
-    "@codemirror/state": "^0.20.0",
-    "@codemirror/view": "^0.20.6",
-    "@floating-ui/core": "^0.7.1",
-    "@floating-ui/dom": "^0.5.1",
-    "@lezer/highlight": "^0.16.0",
-    "@lion/combobox": "^0.10.0",
-    "@lion/core": "^0.22.0",
-    "@lion/listbox": "^0.13.1",
-    "@ngneat/effects": "^1.1.0",
+    "@codemirror/autocomplete": "^6.0.1",
+    "@codemirror/commands": "^6.0.0",
+    "@codemirror/lang-cpp": "^6.0.0",
+    "@codemirror/lang-css": "^6.0.0",
+    "@codemirror/lang-html": "^6.0.0",
+    "@codemirror/lang-java": "^6.0.0",
+    "@codemirror/lang-javascript": "^6.0.0",
+    "@codemirror/lang-json": "^6.0.0",
+    "@codemirror/lang-markdown": "^6.0.0",
+    "@codemirror/lang-php": "^6.0.0",
+    "@codemirror/lang-python": "^6.0.0",
+    "@codemirror/lang-rust": "^6.0.0",
+    "@codemirror/lang-sql": "^6.0.0",
+    "@codemirror/language": "^6.0.0",
+    "@codemirror/lint": "^6.0.0",
+    "@codemirror/search": "^6.0.0",
+    "@codemirror/state": "^6.0.0",
+    "@codemirror/view": "^6.0.0",
+    "@floating-ui/core": "^0.7.3",
+    "@floating-ui/dom": "^0.5.3",
+    "@lezer/highlight": "^1.0.0",
+    "@lion/combobox": "^0.11.0",
+    "@lion/core": "^0.23.0",
+    "@lion/listbox": "^0.14.0",
+    "@ngneat/effects": "^1.1.1",
     "@ngneat/elf": "^2.0.0",
     "@ngneat/elf-devtools": "^1.2.1",
-    "@ngneat/elf-persist-state": "^1.1.1",
+    "@ngneat/elf-persist-state": "^1.1.2",
     "@solid-primitives/i18n": "^1.1.0",
-    "@solid-primitives/immutable": "0.1.0",
-    "@solid-primitives/resize-observer": "^1.1.0",
-    "@thisbeyond/solid-dnd": "0.6.1",
-    "@vanilla-extract/css": "^1.7.0",
+    "@solid-primitives/immutable": "^0.1.0",
+    "@solid-primitives/resize-observer": "^2.0.2",
+    "@thisbeyond/solid-dnd": "^0.6.1",
+    "@vanilla-extract/css": "^1.7.2",
     "@vanilla-extract/dynamic": "^2.0.2",
-    "@vanilla-extract/recipes": "^0.2.4",
+    "@vanilla-extract/recipes": "^0.2.5",
     "@vanilla-extract/sprinkles": "^1.4.1",
     "@vanilla-extract/vite-plugin": "3.1.4",
     "bind-event-listener": "^2.1.0",
@@ -80,13 +80,13 @@
     "idb-keyval": "6.1.0",
     "inter-ui": "^3.19.3",
     "modern-normalize": "^1.1.0",
-    "polished": "4.2.2",
+    "polished": "^4.2.2",
     "rxjs": "^7.5.5",
     "solid-app-router": "^0.3.3",
-    "solid-codemirror": "^1.0.3",
-    "solid-headless": "^0.10.12",
+    "solid-codemirror": "^1.1.0",
+    "solid-headless": "^0.11.0",
     "solid-js": "^1.4.4",
-    "solid-use": "^0.3.4",
+    "solid-use": "^0.4.2",
     "tinykeys": "^1.4.0",
     "workbox-precaching": "^6.5.3",
     "workbox-routing": "^6.5.3",
