+++ apps/codeimage/src/components/Frame/Frame.css.ts
@@ -129,21 +129,6 @@ export const resizeLineDivider = style({
   transform: 'transformY(50%)',
 });
 
-export const squaredBackgroundOverlay = style({
-  backgroundImage: `
-    linear-gradient(45deg,  ${themeVars.dynamicColors.emptySquareBackgroundColor} 25%,transparent 0),
-    linear-gradient(-45deg, ${themeVars.dynamicColors.emptySquareBackgroundColor} 25%,transparent 0),
-    linear-gradient(45deg,transparent 75%, ${themeVars.dynamicColors.emptySquareBackgroundColor} 0),
-    linear-gradient(-45deg,transparent 75%, ${themeVars.dynamicColors.emptySquareBackgroundColor} 0)
-  `,
-  backgroundSize: '20px 20px',
-  backgroundPosition: '0 0,0 10px, 10px -10px, -10px 0',
-  zIndex: 1,
-  position: 'absolute',
-  width: '100%',
-  height: '100%',
-});
-
 export const watermark = style({
   position: 'absolute',
   right: '32px',
