+++ apps/codeimage/src/state/editor/createEditorOptions.ts
@@ -14,7 +14,7 @@ export interface EditorUIOptions {
 function $createEditorOptions() {
   const [, withNotifier, version] = createStoreNotifier();
   const [state, setState] = createStore<EditorUIOptions>({
-    themeId: appEnvironment.defaultState.editor.theme.id,
+    themeId: '',
     showLineNumbers: false,
     fontId: appEnvironment.defaultState.editor.font.id,
     fontWeight: appEnvironment.defaultState.editor.font.types[0].weight,
