+++ packages/config/package.json
@@ -26,40 +26,40 @@
   },
   "dependencies": {
     "@codeimage/highlight": "workspace:*",
-    "@codemirror/autocomplete": "^0.20.1",
-    "@codemirror/commands": "^0.20.0",
-    "@codemirror/language": "^0.20.2",
-    "@codemirror/lint": "^0.20.2",
-    "@codemirror/search": "^0.20.1",
-    "@codemirror/lang-cpp": "^0.20.0",
-    "@codemirror/lang-css": "^0.20.0",
-    "@codemirror/lang-html": "^0.20.0",
-    "@codemirror/lang-java": "^0.20.0",
-    "@codemirror/lang-javascript": "^0.20.0",
-    "@codemirror/lang-json": "^0.20.0",
-    "@codemirror/lang-markdown": "^0.20.1",
-    "@codemirror/lang-php": "^0.20.0",
-    "@codemirror/lang-python": "^0.20.0",
-    "@codemirror/lang-rust": "^0.20.0",
-    "@codemirror/lang-sql": "^0.20.0",
-    "@codemirror/state": "^0.20.0",
-    "@codemirror/view": "^0.20.6",
-    "@codemirror/legacy-modes": "^0.20.0",
-    "material-icon-theme": "^4.17.0",
+    "@codemirror/autocomplete": "^6.0.1",
+    "@codemirror/commands": "^6.0.0",
+    "@codemirror/language": "^6.0.0",
+    "@codemirror/lint": "^6.0.0",
+    "@codemirror/search": "^6.0.0",
+    "@codemirror/lang-cpp": "^6.0.0",
+    "@codemirror/lang-css": "^6.0.0",
+    "@codemirror/lang-html": "^6.0.0",
+    "@codemirror/lang-java": "^6.0.0",
+    "@codemirror/lang-javascript": "^6.0.0",
+    "@codemirror/lang-json": "^6.0.0",
+    "@codemirror/lang-markdown": "^6.0.0",
+    "@codemirror/lang-php": "^6.0.0",
+    "@codemirror/lang-python": "^6.0.0",
+    "@codemirror/lang-rust": "^6.0.0",
+    "@codemirror/lang-sql": "^6.0.0",
+    "@codemirror/state": "^6.0.0",
+    "@codemirror/view": "^6.0.0",
+    "@codemirror/legacy-modes": "^6.0.0",
+    "material-icon-theme": "^4.18.1",
     "solid-js": "^1.4.4"
   },
   "peerDependencies": {
     "@codeimage/highlight": "workspace:*",
-    "@codemirror/state": "^0.20.0",
-    "@codemirror/view": "^0.20.4",
-    "@codemirror/language": "^0.20.0",
+    "@codemirror/state": "^6.0.0",
+    "@codemirror/view": "^6.0.0",
+    "@codemirror/language": "^6.0.0",
     "solid-js": "^1.4.4"
   },
   "devDependencies": {
     "rimraf": "^3.0.2",
     "typescript": "^4.7.3",
-    "vite": "^2.9.8",
-    "vite-plugin-dts": "^1.1.1",
+    "vite": "^2.9.12",
+    "vite-plugin-dts": "^1.2.0",
     "vite-plugin-solid": "^2.2.6"
   }
 }
