+++ packages/locale/package.json
@@ -29,12 +29,12 @@
     "solid-js": "^1.4.4"
   },
   "peerDependencies": {
-    "solid-js": "^1.4.2"
+    "solid-js": "^1.4.4"
   },
   "devDependencies": {
     "rimraf": "^3.0.2",
     "typescript": "^4.7.3",
-    "vite": "^2.9.8",
+    "vite": "^2.9.12",
     "vite-plugin-dts": "^1.1.1",
     "vite-plugin-solid": "^2.2.6"
   }
