+++ packages/ui/package.json
@@ -23,38 +23,38 @@
     "typecheck": "tsc --noEmit --skipLibCheck --project tsconfig.json"
   },
   "peerDependencies": {
-    "@vanilla-extract/css": "^1.7.0",
+    "@vanilla-extract/css": "^1.7.2",
     "@vanilla-extract/sprinkles": "^1.4.1",
-    "@vanilla-extract/recipes": "^0.2.4",
+    "@vanilla-extract/recipes": "^0.2.5",
     "@vanilla-extract/dynamic": "^2.0.2",
-    "solid-headless": "^0.10.12",
+    "solid-headless": "^0.11.0",
     "@solid-aria/interactions": "^0.0.7",
     "@solid-aria/focus": "^0.0.9",
     "@solid-aria/overlays": "^0.0.5",
     "@solid-aria/utils": "^0.0.6",
-    "@solid-primitives/utils": "^2.0.1",
+    "@solid-primitives/utils": "^2.1.0",
     "@solid-primitives/refs": "^0.2.0",
-    "@solid-primitives/props": "^2.1.2",
-    "solid-use": "^0.3.4",
+    "@solid-primitives/props": "^2.1.7",
+    "solid-use": "^0.4.2",
     "solid-js": "^1.4.4",
     "csstype": "^3.1.0",
     "clsx": "^1.1.1",
-    "@floating-ui/dom": "^0.5.1",
-    "@floating-ui/core": "^0.7.1"
+    "@floating-ui/dom": "^0.5.3",
+    "@floating-ui/core": "^0.7.3"
   },
   "dependencies": {
     "@codeimage/locale": "workspace:*",
-    "@vanilla-extract/css": "^1.7.0",
+    "@vanilla-extract/css": "^1.7.2",
     "@vanilla-extract/sprinkles": "^1.4.1",
-    "@vanilla-extract/recipes": "^0.2.4",
+    "@vanilla-extract/recipes": "^0.2.5",
     "@vanilla-extract/dynamic": "^2.0.2",
     "@vanilla-extract/private": "^1.0.3",
-    "solid-headless": "^0.10.12",
-    "solid-use": "^0.3.4",
+    "solid-headless": "^0.11.0",
+    "solid-use": "^0.4.2",
     "csstype": "^3.1.0",
     "clsx": "^1.1.1",
-    "@floating-ui/dom": "^0.5.1",
-    "@floating-ui/core": "^0.7.1",
+    "@floating-ui/dom": "^0.5.3",
+    "@floating-ui/core": "^0.7.3",
     "@solid-aria/interactions": "^0.0.7",
     "@solid-aria/focus": "^0.0.9",
     "@solid-aria/overlays": "^0.0.5",
@@ -71,8 +71,8 @@
     "rimraf": "^3.0.2",
     "solid-js": "^1.4.4",
     "typescript": "^4.7.3",
-    "vite": "^2.9.8",
-    "vite-plugin-dts": "^1.1.1",
+    "vite": "^2.9.12",
+    "vite-plugin-dts": "^1.2.0",
     "vite-plugin-solid": "^2.2.6",
     "tiny-glob": "^0.2.9"
   }
